moved to fuels types
