+++ packages/fuels-core/src/abi_decoder.rs
@@ -5,8 +5,9 @@ use fuel_types::bytes::padded_len_usize;
 use fuels_types::{
     constants::WORD_SIZE,
     errors::CodecError,
-    param_types::{EnumVariants, ParamType},
+    param_types::ParamType,
 };
+use fuels_types::enum_variants::EnumVariants;
 
 #[derive(Debug, Clone)]
 struct DecodeResult {
@@ -323,8 +324,9 @@ fn skip(slice: &[u8], num_bytes: usize) -> Result<&[u8], CodecError> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use fuels_types::{errors::Error, param_types::EnumVariants};
+    use fuels_types::errors::Error;
     use std::vec;
+    use fuels_types::enum_variants::EnumVariants;
 
     #[test]
     fn decode_int() -> Result<(), Error> {
