+++ packages/fuels-core/src/abi_encoder.rs
@@ -1,8 +1,10 @@
 use crate::{
-    pad_string, pad_u16, pad_u32, pad_u8, EnumSelector, EnumVariants, ParamType, StringToken, Token,
+    EnumSelector, pad_string, pad_u16, pad_u32, pad_u8, ParamType, StringToken, Token,
 };
 use fuels_types::{constants::WORD_SIZE, errors::CodecError};
 use itertools::Itertools;
+use fuels_types::enum_variants::EnumVariants;
+
 pub struct ABIEncoder;
 
 #[derive(Debug, Clone)]
@@ -240,11 +242,12 @@ mod tests {
     use crate::utils::first_four_bytes_of_sha256_hash;
     use fuels_types::{
         errors::Error,
-        param_types::{EnumVariants, ParamType},
+        param_types::ParamType,
     };
     use itertools::chain;
     use sha2::{Digest, Sha256};
     use std::slice;
+    use fuels_types::enum_variants::EnumVariants;
 
     const VEC_METADATA_SIZE: usize = 3 * WORD_SIZE;
     const DISCRIMINANT_SIZE: usize = WORD_SIZE;
