+++ packages/fuels-core/src/code_gen.rs
@@ -7,6 +7,5 @@ pub mod docs_gen;
 pub mod function_selector;
 pub mod functions_gen;
 mod resolved_type;
-pub mod utils;
 
 pub use abigen::{extract_and_parse_logs, extract_log_ids_and_data};
