+++ packages/fuels-core/src/code_gen/custom_types.rs
@@ -3,11 +3,9 @@ mod struct_gen;
 mod utils;
 
 pub use enum_gen::expand_custom_enum;
+pub use fuels_types::utils::custom_type_name;
 pub use struct_gen::expand_custom_struct;
-pub use utils::{
-    extract_custom_type_name_from_abi_type_field, param_type_calls, single_param_type_call,
-    Component,
-};
+pub use utils::{param_type_calls, single_param_type_call, Component};
 
 // Doing string -> TokenStream -> string isn't pretty but gives us the opportunity to
 // have a better understanding of the generated code so we consider it ok.
