+++ packages/fuels-types/src/lib.rs
@@ -12,6 +12,7 @@ pub mod constants;
 pub mod errors;
 pub mod param_types;
 pub mod utils;
+pub mod enum_variants;
 
 #[derive(Debug, Clone, Copy, ToString, PartialEq, Eq)]
 #[strum(serialize_all = "lowercase")]
